requirements.txt
# Core dependencies
flask==3.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
requests==2.31.0
python-dotenv==1.0.1

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
markdown==3.5.2

# Data processing (using compatible versions with pre-built wheels)
pandas>=2.0.0
numpy>=1.24.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0